Add Support for using a pipe rather then a socket
