+++ Cargo.toml
@@ -21,6 +21,7 @@ tokio = { version = "1", features = [
   "io-util",
   "macros",
   "rt-multi-thread",
+  "signal",
 ], optional = true }
 
 [dev-dependencies]
