+++ examples/sender.rs
@@ -4,12 +4,9 @@
 //!
 //! The example includes recording a counter, a gauge, and a histogram metric to illustrate the usage of the IPCRecorderBuilder.
 
-use metrics::{counter, gauge, histogram};
-use metrics_ipc_collector::IPCRecorderBuilder;
-
 fn main() {
     // Create an IPCRecorderBuilder and configure the socket path.
-    let builder = IPCRecorderBuilder::default();
+    let builder = metrics_ipc_collector::IPCSocketRecorderBuilder::default();
 
     // Attempt to build the IPC recorder and set it as the global recorder.
     if let Err(e) = builder.build() {
@@ -18,9 +15,9 @@ fn main() {
     }
 
     // Record some example metrics.
-    counter!("example_counter").increment(1);
-    gauge!("example_gauge").set(std::f32::consts::PI);
-    histogram!("example_histogram").record(42.0);
+    metrics::counter!("example_counter").increment(1);
+    metrics::gauge!("example_gauge").set(std::f32::consts::PI);
+    metrics::histogram!("example_histogram").record(42.0);
 
     println!("Metrics recorded and sent to the IPC socket.");
 }
