+++ src/lib.rs
@@ -5,6 +5,12 @@ mod error;
 mod events;
 mod recorder;
 
-pub use collector::IPCCollector;
+pub use collector::{
+    socket::IPCSocketCollector,
+    unnamed_pipe::{IPCPipeCollector, PipeSender},
+};
 pub use error::MetricsError;
-pub use recorder::{IPCRecorder, IPCRecorderBuilder};
+pub use recorder::{
+    socket::{IPCSocketRecorder, IPCSocketRecorderBuilder},
+    unnamed_pipe::IPCPipeRecorder,
+};
